feat: complete Phase 1 stability fixes and clean up stale plans
1.6 Pipeline Tooling:
- Derive EPUB content path from OPF location instead of hardcoded OEBPS
- XHTMLParser now throws on UTF-8/XML parse failures
- MarkerInjector interleaves orphaned markers by epubCharOffset
- SlidingWindowAligner match threshold is now configurable
- TranscribeCommand validates model name before WhisperKit init
- Python transcriber supports --device auto with CUDA detection

1.7 View & UI Model Fixes:
- TransportButton calls configuration.trigger() for VoiceOver/keyboard
- formatHhMm uses truncation instead of rounding up
- TimelineGroup.id includes fractional seconds to prevent collisions
- TranscriptionSegment.id uses integer milliseconds for stability
- Added Equatable/Hashable to AggregatedChapter, M4BBook, Chapter

1.8 watchOS Critical Fixes:
- watchQuickBookmarkTimeoutSeconds is now a computed property
- Voice memos use WCSession.transferFile instead of inline data
- WidgetCenter reload debounced to at most once per 30 seconds
- All haptics gated behind isHapticFeedbackEnabled via playHaptic()

1.9 macOS Critical Fixes:
- process.waitUntilExit() replaced with async continuation
- 300ms Task.sleep replaced with KVO on AVPlayerItem.status
- UserDefaults.standard migrated to AppGroupDefaults.shared

1.10 Performance:
- Cached DateFormatter/ISO8601DateFormatter as static lets
- AlignmentService SQL writes batched into single transaction
- PlaylistView.playlistRows memoized with onChange triggers

Also removes stale .claude/plans/, ALPHA_OVERNIGHT_NOTES.md, neededfixes.md.

+++ Shared/Database/DAOs/PlaybackEventDAO.swift
@@ -11,7 +11,7 @@ struct PlaybackEventDAO {
         endedAt: Date?,
         startPosition: TimeInterval,
         endPosition: TimeInterval?,
-        speed: Float,
+        speed: Double,
         eventType: String,
         source: String?
     ) throws {
@@ -52,7 +52,7 @@ struct PlaybackEvent: Codable, FetchableRecord, MutablePersistableRecord {
     var endedAt: String?
     var startPosition: TimeInterval
     var endPosition: TimeInterval?
-    var speed: Float
+    var speed: Double
     var eventType: String
     var source: String?
